+++ test/_fastify-test-helper.config.js
@@ -1,5 +1,0 @@
-export default {
-	customGetters: {
-		'/shadow-element.js': 'shadow-element.js'
-	}
-};